+++ src/layers/cip/objects/Identity.js
@@ -19,9 +19,16 @@ class Identity extends CIPObject {
 
     let value;
     offset = Decode(dataType, data, offset, val => value = val);
-    const raw = value;
+    // const raw = value;
 
     switch (attribute) {
+      case InstanceAttributeCodes.DeviceType: {
+        value = {
+          code: value,
+          name: DeviceTypeNames[value] || 'Unknown'
+        };
+        break;
+      }
       case InstanceAttributeCodes.Revision: {
         if (Array.isArray(value) && value.length >= 2) {
           value = {
@@ -61,7 +68,7 @@ class Identity extends CIPObject {
         code: attribute,
         name: InstanceAttributeNames[attribute] || 'Unknown',
         value,
-        raw
+        // raw
       });
     }
 
@@ -83,28 +90,6 @@ class Identity extends CIPObject {
       return accum;
     }, []);
 
-    // const attributes = [
-    //   InstanceAttributeCodes.VendorID,
-    //   InstanceAttributeCodes.DeviceType,
-    //   InstanceAttributeCodes.ProductCode,
-    //   InstanceAttributeCodes.Revision,
-    //   InstanceAttributeCodes.Status,
-    //   InstanceAttributeCodes.SerialNumber,
-    //   InstanceAttributeCodes.ProductName
-    // ].reduce((accum, attribute) => {
-    //   offset = this.DecodeInstanceAttribute(attribute, buffer, offset, val => accum[attribute] = val);
-    //   return accum;
-    // }, {});
-
-    // const attributes = {};
-    // offset = this.DecodeInstanceAttribute(InstanceAttributeCodes.VendorID, buffer, offset, val => attributes.vendorID = val);
-    // offset = this.DecodeInstanceAttribute(InstanceAttributeCodes.DeviceType, buffer, offset, val => attributes.deviceType = val);
-    // offset = this.DecodeInstanceAttribute(InstanceAttributeCodes.ProductCode, buffer, offset, val => attributes.productCode = val);
-    // offset = this.DecodeInstanceAttribute(InstanceAttributeCodes.Revision, buffer, offset, val => attributes.revision = val);
-    // offset = this.DecodeInstanceAttribute(InstanceAttributeCodes.Status, buffer, offset, val => attributes.status = val);
-    // offset = this.DecodeInstanceAttribute(InstanceAttributeCodes.SerialNumber, buffer, offset, val => attributes.serialNumber = val);
-    // offset = this.DecodeInstanceAttribute(InstanceAttributeCodes.ProductName, buffer, offset, val => attributes.productName = val);
-
     if (typeof cb === 'function') {
       cb(attributes);
     }
@@ -113,27 +98,12 @@ class Identity extends CIPObject {
   }
 
 
-  // static ParseInstanceAttributeState(buffer, offset, cb) {
-  //   const code = buffer.readUInt8(offset); offset += 1;
-  //   const state = {
-  //     code,
-  //     description: InstanceStateDescriptions[code] || 'unknown'
-  //   };
-
-  //   if (typeof cb === 'function') {
-  //     cb(state);
-  //   }
-
-  //   return offset;
-  // }
-
-
   static get Services() {
     return ClassServices;
   }
 
 
-  static DecodeGetAttributesAll(data, offset, cb) {
+  static DecodeGetClassAttributesAll(data, offset, cb) {
     const info = {};
     info.data = data;
     const length = data.length;
@@ -302,6 +272,98 @@ const ExtendedDeviceStatusDescriptions = {
   0b0101: 'Major fault - either bit 10 or bit 11 is true (1)',
   0b0110: 'At least one I/O connection in run mode',
   0b0111: 'At least one I/O connection established, all in idle mode'
+};
+
+/** CIP Vol 1, Table 6-7.1 */
+const DeviceTypeCodes = {
+  ACDrives: 0x02,
+  ModbusDevice: 0x28,
+  CIPMotionDrive: 0x25,
+  CommunicationsAdapter: 0x0C,
+  CompoNetRepeater: 0x26,
+  Contactor: 0x15,
+  ControlNetPhysicalLayerComponent: 0x32,
+  ProgrammableLogicController: 0x0E,
+  DCDrives: 0x13,
+  DCPowerGenerator: 0x1F,
+  Encoder: 0x22,
+  FluidFlowController: 0x24,
+  GeneralPurposeDiscreteIO: 0x07,
+  GenericDevice: 0x00,
+  HumanMachineInterface: 0x18,
+  InductiveProximitySwitch: 0x05,
+  LimitSwitch: 0x04,
+  MassFlowController: 0x1A,
+  MassFlowControllerEnhanced: 0x27,
+  MotorOverload: 0x03,
+  MotorStarter: 0x16,
+  PhotoelectricSensor: 0x06,
+  PneumaticValve: 0x1B,
+  PositionController: 0x10,
+  ProcessControlValve: 0x1D,
+  ResidualGasAnalyzer: 0x1E,
+  Resolver: 0x09,
+  RFPowerGenerator: 0x20,
+  SafetyDiscreteIODevice: 0x23,
+  SoftstartStarter: 0x17,
+  TurbomolecularVacuumPump: 0x21,
+  VacuumPressureGauge: 0x1C,
+
+  /** OBSOLETE DEVICE PROFILES */
+  OBSOLETEControlStation: 0x01,
+  OBSOLETEEncoder: 0x08,
+  OBSOLETEGeneralPurposeAnalogIO: 0x0A,
+  OBSOLETEBarcodeScanner: 0x0D,
+  OBSOLETEWeightScale: 0x11,
+  OBSOLETEMessageDisplay: 0x12,
+  OBSOLETEServoDrives: 0x14,
+  OBSOLETEPneumaticValve: 0x19
+};
+
+/** CIP Vol 1, Table 6-7.1 */
+const DeviceTypeNames = {
+  [DeviceTypeCodes.ACDrives]: 'AC Drives',
+  [DeviceTypeCodes.ModbusDevice]: 'Modbus Device',
+  [DeviceTypeCodes.CIPMotionDrive]: 'CIP Motion Drive',
+  [DeviceTypeCodes.CommunicationsAdapter]: 'Communications Adapter',
+  [DeviceTypeCodes.CompoNetRepeater]: 'CompoNet Repeater',
+  [DeviceTypeCodes.Contactor]: 'Contactor',
+  [DeviceTypeCodes.ControlNetPhysicalLayerComponent]: 'ControlNet Physical Layer Component',
+  [DeviceTypeCodes.ProgrammableLogicController]: 'Programmable Logic Controller',
+  [DeviceTypeCodes.DCDrives]: 'DC Drives',
+  [DeviceTypeCodes.DCPowerGenerator]: 'DC Power Generator',
+  [DeviceTypeCodes.Encoder]: 'Encoder',
+  [DeviceTypeCodes.FluidFlowController]: 'Fluid Flow Controller',
+  [DeviceTypeCodes.GeneralPurposeDiscreteIO]: 'General Purpose Discrete I/O',
+  [DeviceTypeCodes.GenericDevice]: 'Generic Device',
+  [DeviceTypeCodes.HumanMachineInterface]: 'Human-Machine Interface',
+  [DeviceTypeCodes.InductiveProximitySwitch]: 'Inductive Proximity Switch',
+  [DeviceTypeCodes.LimitSwitch]: 'Limit Switch',
+  [DeviceTypeCodes.MassFlowController]: 'Mass Flow Controller',
+  [DeviceTypeCodes.MassFlowControllerEnhanced]: 'Mass Flow Controller, Enhanced',
+  [DeviceTypeCodes.MotorOverload]: 'Motor Overload',
+  [DeviceTypeCodes.MotorStarter]: 'Motor Starter',
+  [DeviceTypeCodes.PhotoelectricSensor]: 'Photoelectric Sensor',
+  [DeviceTypeCodes.PneumaticValve]: 'Pneumatic Valve(s)',
+  [DeviceTypeCodes.PositionController]: 'Position Controller',
+  [DeviceTypeCodes.ProcessControlValve]: 'Process Control Valve',
+  [DeviceTypeCodes.ResidualGasAnalyzer]: 'Residual Gas Analyzer',
+  [DeviceTypeCodes.Resolver]: 'Resolver',
+  [DeviceTypeCodes.RFPowerGenerator]: 'RF Power Generator',
+  [DeviceTypeCodes.SafetyDiscreteIODevice]: 'Safety Discrete I/O Device',
+  [DeviceTypeCodes.SoftstartStarter]: 'Softstart Starter',
+  [DeviceTypeCodes.TurbomolecularVacuumPump]: 'Turbomolecular Vacuum Pump',
+  [DeviceTypeCodes.VacuumPressureGauge]: 'Vacuum Pressure Gauge',
+
+  /** OBSOLETE DEVICE PROFILES */
+  [DeviceTypeCodes.OBSOLETEControlStation]: '(OBSOLETE) Control Station',
+  [DeviceTypeCodes.OBSOLETEEncoder]: '(OBSOLETE) Encoder',
+  [DeviceTypeCodes.OBSOLETEGeneralPurposeAnalogIO]: '(OBSOLETE) General Purpose Analog I/O',
+  [DeviceTypeCodes.OBSOLETEBarcodeScanner]: '(OBSOLETE) Barcode Scanner',
+  [DeviceTypeCodes.OBSOLETEWeightScale]: '(OBSOLETE) Weight Scale',
+  [DeviceTypeCodes.OBSOLETEMessageDisplay]: '(OBSOLETE) Message Display',
+  [DeviceTypeCodes.OBSOLETEServoDrives]: '(OBSOLETE) Servo Drives',
+  [DeviceTypeCodes.OBSOLETEPneumaticValve]: '(OBSOLETE) Pneumatic Valve(s)'
 };